+++ models/projectbond.js
@@ -32,7 +32,7 @@ module.exports = (sequelize, DataTypes) => {
       percent: DataTypes.DOUBLE,
       institution_type:DataTypes.STRING,
       phone: DataTypes.STRING,
-      amount: DataTypes.INTEGER, //fix to double
+      amount: DataTypes.DOUBLE, //fix to double
       remark: DataTypes.TEXT,
       revision_no: DataTypes.INTEGER,
     },
